feat: handle deletion responses
- refactor: split up Table implementaions into multiple modules
